+++ packages/moonstone/VideoPlayer/VideoPlayer.js
@@ -18,8 +18,8 @@ import ilib from '@enact/i18n';
 import DurationFmt from '@enact/i18n/ilib/lib/DurationFmt';
 import {toUpperCase} from '@enact/i18n/util';
 import Spotlight from '@enact/spotlight';
-import {SpotlightContainerDecorator, spotlightDefaultClass} from '@enact/spotlight/SpotlightContainerDecorator';
-import {Spottable, spottableClass} from '@enact/spotlight/Spottable';
+import {SpotlightContainerDecorator} from '@enact/spotlight/SpotlightContainerDecorator';
+import {Spottable} from '@enact/spotlight/Spottable';
 import Announce from '@enact/ui/AnnounceDecorator/Announce';
 import ComponentOverride from '@enact/ui/ComponentOverride';
 import {FloatingLayerDecorator} from '@enact/ui/FloatingLayer';
@@ -226,6 +226,7 @@ const VideoPlayerBase = class extends React.Component {
 		 * * `onRewind` - Called when the media is rewound via a key event
 		 * * `onToggleMore` - Called when the more components are hidden or shown
 		 * * `paused` - `true` when the media is paused
+		 * * `spotlightId` - The spotlight container Id for the media controls
 		 * * `spotlightDisabled` - `true` when spotlight is disabled for the media controls
 		 * * `visible` - `true` when the media controls should be displayed
 		 *
@@ -490,6 +491,7 @@ const VideoPlayerBase = class extends React.Component {
 		 *
 		 * @type {String}
 		 * @public
+		 * @default 'videoPlayer'
 		 */
 		spotlightId: PropTypes.string,
 
@@ -594,6 +596,7 @@ const VideoPlayerBase = class extends React.Component {
 			slowForward: ['1/4', '1/2'],
 			slowRewind: ['-1/2', '-1']
 		},
+		spotlightId: 'videoPlayer',
 		titleHideDelay: 5000,
 		videoComponent: Media
 	}
@@ -611,6 +614,7 @@ const VideoPlayerBase = class extends React.Component {
 		this.id = this.generateId();
 		this.selectPlaybackRates('fastForward');
 		this.sliderKnobProportion = 0;
+		this.mediaControlsSpotlightId = props.spotlightId + '_mediaControls';
 
 		this.initI18n();
 
@@ -699,6 +703,10 @@ const VideoPlayerBase = class extends React.Component {
 			this.context.closeAllFloatingLayers();
 		}
 
+		if (this.props.spotlightId !== prevProps.spotlightId) {
+			this.mediaControlsSpotlightId = this.props.spotlightId + '_mediaControls';
+		}
+
 		if (!this.state.mediaControlsVisible && prevState.mediaControlsVisible) {
 			forwardControlsAvailable({available: false}, this.props);
 			this.stopAutoCloseTimeout();
@@ -716,7 +724,7 @@ const VideoPlayerBase = class extends React.Component {
 				const current = Spotlight.getCurrent();
 				if (!current || this.player.contains(current)) {
 					// Set focus within media controls when they become visible.
-					this.focusDefaultMediaControl();
+					Spotlight.focus(this.mediaControlsSpotlightId);
 				}
 			}
 		}
@@ -1532,25 +1540,6 @@ const VideoPlayerBase = class extends React.Component {
 		this.hideControls
 	)
 
-	/**
-	 * Check for elements with the spotlightDefaultClass, in the following location order:
-	 * left components, right components, media controls or more controls (depending on which is
-	 * available)
-	 *
-	 * @returns {Node|false} The focused control or `false` if nothing is found.
-	 * @private
-	 */
-	focusDefaultMediaControl = () => {
-		const defaultSpottable = `.${spotlightDefaultClass}.${spottableClass}`;
-		const defaultControl =
-			this.player.querySelector(
-				`.${css.leftComponents} ${defaultSpottable}, .${css.rightComponents} ${defaultSpottable}`
-			) ||
-			this.player.querySelector(`[data-media-controls] ${defaultSpottable}`);
-
-		return defaultControl ? Spotlight.focus(defaultControl) : false;
-	}
-
 	//
 	// Player Interaction events
 	//
@@ -1635,7 +1624,7 @@ const VideoPlayerBase = class extends React.Component {
 		} else if (is('down', ev.keyCode)) {
 			Spotlight.setPointerMode(false);
 
-			if (this.focusDefaultMediaControl()) {
+			if (Spotlight.focus(this.mediaControlsSpotlightId)) {
 				stopImmediate(ev);
 			}
 		} else if (is('up', ev.keyCode)) {
@@ -1878,6 +1867,7 @@ const VideoPlayerBase = class extends React.Component {
 								onRewind={this.handleRewind}
 								onToggleMore={this.handleToggleMore}
 								paused={this.state.paused}
+								spotlightId={this.mediaControlsSpotlightId}
 								spotlightDisabled={!this.state.mediaControlsVisible || spotlightDisabled}
 								visible={this.state.mediaControlsVisible}
 							/>